// apps/api/migrations/snapshots/b0854eed46485075d8b9797eec1b314dedbe8d59cdd3e17b26f55ecb6c3d346f/contract.d.ts
export type CodecTypes = PgTypes & PostgisTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes> &
  PostgisQueryOperationTypes<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Country: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly code: CodecTypes['pg/text@1']['output'];
    };
    readonly Property: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly propertyType:
        | 'ESTATE'
        | 'COMPOUND'
        | 'BUILDING'
        | 'HOUSE'
        | 'APARTMENT'
        | 'HOSTEL'
        | 'SHOP'
        | 'OFFICE'
        | 'WAREHOUSE'
        | 'LAND';
      readonly countryId: CodecTypes['pg/text@1']['output'];
      readonly stateId: CodecTypes['pg/text@1']['output'];
      readonly area: CodecTypes['pg/text@1']['output'];
      readonly city: CodecTypes['pg/text@1']['output'] | null;
      readonly address: CodecTypes['pg/text@1']['output'] | null;
      readonly landmark: CodecTypes['pg/text@1']['output'] | null;
      readonly latitude: CodecTypes['pg/float8@1']['output'] | null;
      readonly longitude: CodecTypes['pg/float8@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly PropertyRelationship: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly parentPropertyId: CodecTypes['pg/text@1']['output'];
      readonly childPropertyId: CodecTypes['pg/text@1']['output'];
      readonly relationshipType: 'CONTAINS';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly State: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly code: CodecTypes['pg/text@1']['output'];
      readonly countryId: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Country: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly code: CodecTypes['pg/text@1']['input'];
    };
    readonly Property: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly propertyType:
        | 'ESTATE'
        | 'COMPOUND'
        | 'BUILDING'
        | 'HOUSE'
        | 'APARTMENT'
        | 'HOSTEL'
        | 'SHOP'
        | 'OFFICE'
        | 'WAREHOUSE'
        | 'LAND';
      readonly countryId: CodecTypes['pg/text@1']['input'];
      readonly stateId: CodecTypes['pg/text@1']['input'];
      readonly area: CodecTypes['pg/text@1']['input'];
      readonly city: CodecTypes['pg/text@1']['input'] | null;
      readonly address: CodecTypes['pg/text@1']['input'] | null;
      readonly landmark: CodecTypes['pg/text@1']['input'] | null;
      readonly latitude: CodecTypes['pg/float8@1']['input'] | null;
      readonly longitude: CodecTypes['pg/float8@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly PropertyRelationship: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly parentPropertyId: CodecTypes['pg/text@1']['input'];
      readonly childPropertyId: CodecTypes['pg/text@1']['input'];
      readonly relationshipType: 'CONTAINS';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly State: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly code: CodecTypes['pg/text@1']['input'];
      readonly countryId: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly country: {
      readonly code: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly property: {
      readonly address: CodecTypes['pg/text@1']['output'] | null;
      readonly area: CodecTypes['pg/text@1']['output'];
      readonly city: CodecTypes['pg/text@1']['output'] | null;
      readonly countryId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly landmark: CodecTypes['pg/text@1']['output'] | null;
      readonly latitude: CodecTypes['pg/float8@1']['output'] | null;
      readonly longitude: CodecTypes['pg/float8@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly propertyType:
        | 'ESTATE'
        | 'COMPOUND'
        | 'BUILDING'
        | 'HOUSE'
        | 'APARTMENT'
        | 'HOSTEL'
        | 'SHOP'
        | 'OFFICE'
        | 'WAREHOUSE'
        | 'LAND';
      readonly stateId: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly propertyRelationship: {
      readonly childPropertyId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly parentPropertyId: CodecTypes['pg/text@1']['output'];
      readonly relationshipType: 'CONTAINS';
    };
    readonly state: {
      readonly code: CodecTypes['pg/text@1']['output'];
      readonly countryId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly country: {
      readonly code: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly property: {
      readonly address: CodecTypes['pg/text@1']['input'] | null;
      readonly area: CodecTypes['pg/text@1']['input'];
      readonly city: CodecTypes['pg/text@1']['input'] | null;
      readonly countryId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly landmark: CodecTypes['pg/text@1']['input'] | null;
      readonly latitude: CodecTypes['pg/float8@1']['input'] | null;
      readonly longitude: CodecTypes['pg/float8@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly propertyType:
        | 'ESTATE'
        | 'COMPOUND'
        | 'BUILDING'
        | 'HOUSE'
        | 'APARTMENT'
        | 'HOSTEL'
        | 'SHOP'
        | 'OFFICE'
        | 'WAREHOUSE'
        | 'LAND';
      readonly stateId: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly propertyRelationship: {
      readonly childPropertyId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly parentPropertyId: CodecTypes['pg/text@1']['input'];
      readonly relationshipType: 'CONTAINS';
    };
    readonly state: {
      readonly code: CodecTypes['pg/text@1']['input'];
      readonly countryId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;